+++ R/quorum.R
@@ -12,7 +12,7 @@
 #'   share of total votes if less than 1, otherwise as number of votes. Must be
 #'   greater than 0. Uses [reached_quorum_total()].
 #'
-#' @details There's a difference in how the functions work. With `quroum_any`,
+#' @details There's a difference in how the functions work. With `quorum_any`,
 #'   \emph{at least one} quorum must be reached. With `quorum_all` \emph{all}
 #'   (i.e. both) quorums must be reached. If you only use one parameter,
 #'   `quorum_any()` and `quorum_all()` are identical.
@@ -21,6 +21,8 @@
 #'   a boolean vector with length equal to the number of lists/parties
 #'   (`votes_matrix` rows). The vector shows whether a party has reached any/all
 #'   quorums.
+#'
+#' @seealso [apply_quorum()] for standalone quorum calculations
 #'
 #' @examples
 #' votes_matrix = matrix(c(502, 55, 80, 10, 104, 55, 0, 1), ncol = 2)
@@ -74,17 +76,17 @@ create_quorum_function_list = function(type, any_district, total) {
         })
     }
 
-    attributes(quorum_funcs)$type <- type
+    attributes(quorum_funcs)[["type"]] <- type
     return(quorum_funcs)
 }
 
-#' Check if lists/parties have reached the quorum for all votes
+#' Check if parties reached the quorum for all votes
 #'
-#' Base implementation, used by [quorum_any()] and [quorum_all()].
+#' Base implementation, used by \code{\link[=quorum_functions]{quorum_functions}}.
 #'
 #' @param votes_matrix votes matrix
 #' @param quorum_total Vote threshold a party must reach for all votes cast.
-#'                     Used as quota of total votes if less than 1, otherwise
+#'                     Used as fraction of total votes if less than 1, otherwise
 #'                     as number of votes. Must be greater than 0.
 #'
 #' @returns Logical vector with length equal to the number of lists/parties (`votes_matrix`
@@ -106,14 +108,14 @@ reached_quorum_total = function(votes_matrix, quorum_total) {
     return(passed_total_quor)
 }
 
-#' Check if lists/parties have reached a quorum in at least one district
+#' Check if parties reached a quorum in at least one district
 #'
-#' Base implementation, used by [quorum_any()] and [quorum_all()].
+#' Base implementation, used by \code{\link[=quorum_functions]{quorum_functions}}.
 #'
 #' @param votes_matrix votes matrix
 #' @param quorum_districts Vote threshold a party must reach in \emph{at least}
-#'                         one district. Used as quota of total votes within a
-#'                         district if less than 1 otherwise as number of votes.
+#'                         one district. Used as fraction of total votes within a
+#'                         district if less than 1, otherwise as number of votes.
 #'                         Must be greater than 0.
 #'
 #' @inherit reached_quorum_total return
@@ -144,16 +146,14 @@ reached_quorum_any_district = function(votes_matrix, quorum_districts) {
 #' called within [biproporz()]. There's generally no need to call it
 #' directly.
 #'
-#' @seealso [quorum_all()], [quorum_any()] to create a list of quorum functions.
+#' @seealso \code{\link[=quorum_functions]{quorum_functions}} to create a list of quorum
+#'   functions.
 #'
 #' @inherit reached_quorum_total return
 #' @keywords internal
 reached_quorums = function(votes_matrix, quorum_funcs) {
     assert(is.matrix(votes_matrix))
-    if(!is.list(quorum_funcs) || !is.function(quorum_funcs[[1]])) {
-        stop("`", deparse(substitute(quorum_funcs)),
-             "` is not a list of functions.", call. = FALSE)
-    }
+    assert(is_quorum_function_list(quorum_funcs))
 
     # list of vector whether quorum was reached for each party
     has_reached_quorum = lapply(quorum_funcs, function(qf) {
@@ -162,19 +162,56 @@ reached_quorums = function(votes_matrix, quorum_funcs) {
 
     if(length(quorum_funcs) == 1) {
         return(quorum_funcs[[1]](votes_matrix))
-    } else if(is.null(attributes(quorum_funcs)$type)) {
-        stop("type must be set as list attribute.", call. = FALSE)
     }
 
     quorum_matrix = do.call(cbind, has_reached_quorum)
-    if(attributes(quorum_funcs)$type == "ALL") {
+    if(attributes(quorum_funcs)[["type"]] == "ALL") {
         quorum_bool = apply(quorum_matrix, 1, all)
-    } else if(attributes(quorum_funcs)$type == "ANY") {
+    } else if(attributes(quorum_funcs)[["type"]] == "ANY") {
         quorum_bool = apply(quorum_matrix, 1, any)
-    } else {
-        stop("Unknown type `", attributes(quorum_funcs)$type, "`.", call. = FALSE)
     }
     return(quorum_bool)
+}
+
+#' Apply quorum to votes vector or matrix
+#'
+#' This quorum calculation implementation is called within [proporz()], [biproporz()] and
+#' related functions. Generally, there's no need to call `apply_quorum` directly.
+#'
+#' @param votes votes vector or votes matrix
+#' @param quorum Depending on `votes`:
+#'   * For a vector: Vote threshold a party must reach. Used as fraction of total
+#'   votes if less than 1 otherwise as number of votes.
+#'   * For a matrix: List of quorum functions (created with \code{\link[=quorum_functions]{quorum_functions}})
+#'   or a logical vector with the same length as the number of `votes` rows.
+#'
+#' @return Vector or matrix with same dimension as `votes`. Parties that failed to reach the
+#'   specified quorum have their votes set to zero.
+#'
+#' @seealso \code{\link[=quorum_functions]{quorum_functions}} for more matrix examples.
+#'
+#' @export
+#'
+#' @examples
+#' # vector
+#' (votes = c(81, 9, 10))
+#'
+#' apply_quorum(votes, 10)
+#'
+#' apply_quorum(votes, .11)
+#'
+#' # matrix
+#' (votes_matrix = matrix(c(91, 9, 199, 1), nrow = 2))
+#'
+#' apply_quorum(votes_matrix, quorum_all(total = 0.1))
+#'
+#' apply_quorum(votes_matrix, c(FALSE, TRUE))
+apply_quorum = function(votes, quorum) {
+    assert(is.matrix(votes) || is.atomic(votes))
+    if(is.matrix(votes)) {
+        return(apply_quorum_matrix(votes, quorum))
+    }
+    apply_quorum_vector(votes, quorum)
 }
 
 # quorum has to be a list of functions created by quorum_all/any because
@@ -183,13 +220,13 @@ reached_quorums = function(votes_matrix, quorum_funcs) {
 # and hopefully reduce misunderstandings
 apply_quorum_matrix = function(votes_matrix, quorum) {
     assert(is.matrix(votes_matrix))
+    check_quorum_param_matrix(quorum)
+
     if(is.list(quorum)) {
         quorum_bool = reached_quorums(votes_matrix, quorum)
     } else if(is.vector(quorum) && is.logical(quorum)) {
         assert(length(quorum) == nrow(votes_matrix))
         quorum_bool = quorum
-    } else {
-        stop("Cannot parse quorum function or vector.", call. = FALSE)
     }
 
     if(any(!quorum_bool)) {
@@ -199,19 +236,41 @@ apply_quorum_matrix = function(votes_matrix, quorum) {
     return(votes_matrix)
 }
 
+check_quorum_param_matrix = function(quorum) {
+    if(!(is_quorum_function_list(quorum) || (is.atomic(quorum) && is.logical(quorum)))) {
+        stop("Quorum parameter must be a logical vector or a list of quorum functions (see ?quorum_functions)",
+             call. = FALSE)
+    }
+    invisible(TRUE)
+}
+
 # quorum for single vector for proporz() methods
 apply_quorum_vector = function(votes_vector, quorum) {
-    check_votes_vector(votes_vector)
-    assert(length(quorum) == 1 && is.numeric(quorum) && quorum >= 0)
+    check_votes_vector(votes_vector, deparse(substitute(votes)))
+    check_quorum_param_vector(quorum)
 
     if(quorum < 1) {
         quorum = ceiling(sum(votes_vector)*quorum)
     }
 
     if(all(votes_vector < quorum)) {
-        stop("No party reached the quorum.", call. = FALSE)
+        stop("No party reached the quorum", call. = FALSE)
     }
 
     votes_vector[votes_vector < quorum] <- 0
     return(votes_vector)
+}
+
+check_quorum_param_vector = function(quorum) {
+    if(!(length(quorum) == 1 && is.numeric(quorum) && quorum >= 0)) {
+        stop("Quorum parameter must be a single number >= 0",
+             call. = FALSE)
+    }
+    invisible(TRUE)
+}
+
+is_quorum_function_list = function(q) {
+    if(!is.list(q) || is.data.frame(q)) return(FALSE)
+    if(length(setdiff(c("ANY", "ALL"), attributes(q)[["type"]])) != 1) return(FALSE)
+    all(sapply(q, is.function))
 }